ohrrpgce: Fix ptshapes list to really be consistent (ie. (NFRAMES, WIDTH, HEIGHT))
Sorry TMC, this contradicts what I said before. I just failed to think about and communicate my
intent, which is to firstly to be able to manipulate the data with a minimum of munging
(here when we want to memmap a .pt? file, we still need to halve the Y size. I figure the
 wrapper can reasonably hide this.)